% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayessource-package.R
\docType{package}
\name{bayessource-package}
\alias{bayessource-package}
\title{Package {bayessource}}
\description{
Bayesian approach to evaluate whether two sets of multivariate observations come from the same source.
}
\details{
The observations are assumed to be generated by a hierarchical Normal-Inverted Wishart distribution.
The hyperparameters can be fitted using additional background data, covering samples from multiple sources.

The package implements a Gibbs sampler to sample from the posteriors, and the computation of the marginal likelihood follows Chib (1995).
The Bayes factor can also be computed as a ratio of two marginal likelihoods.
}
\section{Normal-Inverse-Wishart model}{


Described in \insertCite{Bozza2008Probabilistic}{bayessource}.

Observation level:
\itemize{
\item \deqn{X_{ij} \sim N_p(\theta_i, W_i)}{X_{ij} ~ N_p(theta_i, W_i)} (i = source, j = items from source)
}

Group level:
\itemize{
\item \deqn{\theta_i \sim N_p(\mu, B)}{theta_i ~ N_p(\mu, B)}
\item \deqn{W_i \sim IW_p(\nu_w, U)}{W_i ~ IW_p(nw, U)}
}

Hyperparameters:
\itemize{
\item \deqn{B, U, \nu_w, \mu}{B, U, nw, \mu}
}

Posterior samples of \eqn{\theta}{theta}, \eqn{W^{(-1)}} can be generated with a Gibbs sampler.
}

\references{
\insertAllCited{}
}
\seealso{
Other core functions: 
\code{\link{get_minimum_nw_IW}()},
\code{\link{make_priors_and_init}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{marginalLikelihood}()},
\code{\link{mcmc_postproc}()},
\code{\link{samesource_C}()},
\code{\link{two.level.multivariate.calculate.UC}()}
}
\concept{core functions}
