% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayessource-package.R
\docType{package}
\name{bayessource}
\alias{bayessource}
\alias{bayessource-package}
\title{bayessource}
\description{
Bayesian approach to evaluate whether two datasets come from the same population.
}
