% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chol2inv}
\alias{chol2inv}
\title{Compute the inverse from the upper Cholesky factor}
\usage{
chol2inv(U_chol)
}
\description{
Compute the inverse from the upper Cholesky factor
}
\seealso{
Other C++ functions: 
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{isCholeskyOn}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other math functions: 
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_pd}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()}
}
\concept{C++ functions}
\concept{math functions}
\keyword{internal}
