% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diwishart_inverse}
\alias{diwishart_inverse}
\title{Inverted Wishart density from the inverse (faster).}
\usage{
diwishart_inverse(X_inv, df, Sigma, logd = FALSE, is_chol = FALSE)
}
\arguments{
\item{X_inv}{inverse of X (the observation)}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{logd}{if TRUE, return the log-density}

\item{is_chol}{if TRUE, Sigma and X_inv are the upper Cholesky factors of Sigma and X.inv}
}
\description{
Computes the density of an Inverted Wishart (df, Sigma) in X, by supplying (X^(-1), df, Sigma) rather than (X, df, Sigma).
Avoids a matrix inversion.
}
\details{
Computes the pdf p_X(x) by knowing x^(-1)
}
\section{Inverted Wishart parametrization (Press)}{


Uses Press[Press2012] parametrization.

\deqn{X ~ IW(v, S)}
with \eqn{S = pxp} matrix, \eqn{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = S / (v - 2(p + 1))}
}

\references{
[Press2012] S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.
}
\seealso{
\code{\link{diwishart}}, \code{\link{dwishart}}
}
