% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diwishart_inverse_R.R
\name{diwishart_inverse_R}
\alias{diwishart_inverse_R}
\title{Inverted Wishart density from the inverse.}
\usage{
diwishart_inverse_R(X.inv, df, Sigma, log = FALSE, is.chol = FALSE)
}
\arguments{
\item{X.inv}{inverse of X (the observation)}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{log}{if TRUE, return the log-density}

\item{is.chol}{if TRUE, Sigma and X.inv are the upper Cholesky factors of Sigma and X.inv}
}
\value{
the density in X
}
\description{
Computes the density of an Inverted Wishart (df, Sigma) in X, by supplying (X^(-1), df, Sigma) rather than (X, df, Sigma).
Avoids a matrix inversion.
}
\details{
Computes the pdf p_X(x) by knowing x^(-1)
}
\section{Inverted Wishart parametrization (Press)}{


Uses Press\link{Press2012} parametrization.

\deqn{X ~ IW(v, S)}
with \eqn{S = pxp} matrix, \eqn{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = S / (v - 2(p + 1))}
}

\references{
\link{Press2012} S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.
}
\seealso{
\code{\link{diwishart_inverse}}, \code{\link{diwishart}}
}
\keyword{internal}
