% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Multivariate normal density. Assumes symmetry.}
\usage{
dmvnorm(x, mean, Cov, logd = FALSE, is_chol = FALSE)
}
\arguments{
\item{x}{the observation (nxp)}

\item{mean}{mean vector (row vector, 1xp)}

\item{Cov}{covariance matrix (pxp)}

\item{logd}{if TRUE, return the log-density}

\item{is_chol}{if TRUE, Cov is the upper Cholesky factor of Cov}
}
\value{
the density in x (nx1)
}
\description{
Faster than \code{dmvnorm} in package \pkg{mvtnorm}. Implemented in C.
}
