% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwishart.R
\name{dwishart}
\alias{dwishart}
\title{Wishart density.}
\usage{
dwishart(X, df, Sigma, log = FALSE, is.chol = FALSE)
}
\arguments{
\item{X}{the observation}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{log}{if TRUE, return the log-density}

\item{is.chol}{if TRUE, Sigma and X are the upper Cholesky factors of Sigma and X}
}
\value{
the density in X
}
\description{
Density of X ~ Wishart(df, Sigma).
}
\section{Wishart parametrization}{


Uses \insertCite{Press2012Applied,Anderson2003introduction}{bayessource} parametrization.

\deqn{X ~ W(v, S)}
with \eqn{S = pxp} matrix, \eqn{v >= p} (the degrees of freedom).

Then:
\deqn{E[X] = v * S}
}

\references{
\insertAllCited{}
}
