% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_Cholesky_from_Cholesky}
\alias{inv_Cholesky_from_Cholesky}
\title{Upper Cholesky factor of inverse from upper Cholesky factor}
\usage{
inv_Cholesky_from_Cholesky(U)
}
\description{
If \eqn{A = U' U}, compute \eqn{V} where \eqn{A^{(-1)} = V' V}
}
\keyword{internal}
