% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isCholeskyOn}
\alias{isCholeskyOn}
\title{Check whether Cholesky factorization is used.}
\usage{
isCholeskyOn()
}
\description{
Check whether Cholesky factorization is used.
}
\seealso{
Other C++ functions: 
\code{\link{chol2inv}()},
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}
}
\concept{C++ functions}
\keyword{internal}
