% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldet_from_Cholesky}
\alias{ldet_from_Cholesky}
\title{log-determinant from Cholesky factor}
\usage{
ldet_from_Cholesky(T_chol)
}
\description{
If \eqn{A = U' U}, compute log(det(A)) from U
}
\keyword{internal}
