% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logCummeanExp}
\alias{logCummeanExp}
\title{Computes log( cummean_i( exp(v[i]) ) ) in a stable way.}
\usage{
logCummeanExp(v)
}
\description{
Computes log( cummean_i( exp(v[i]) ) ) in a stable way.
}
\keyword{internal}
