% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logCumsumExp}
\alias{logCumsumExp}
\title{Computes log( cumsum_i( exp(v_i) )) in a stable way.}
\usage{
logCumsumExp(v)
}
\description{
Computes log( cumsum_i( exp(v_i) )) in a stable way.
}
\seealso{
Other C++ functions: 
\code{\link{chol2inv}()},
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{isCholeskyOn}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other math functions: 
\code{\link{chol2inv}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_pd}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()}
}
\concept{C++ functions}
\concept{math functions}
\keyword{internal}
