% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logSumExp}
\alias{logSumExp}
\title{Computes \eqn{log( sum_i( exp(v[i] )) )} in a stable way.}
\usage{
logSumExp(v)
}
\description{
Computes \eqn{log( sum_i( exp(v[i] )) )} in a stable way.
}
\keyword{internal}
