% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logSumExpMean}
\alias{logSumExpMean}
\title{Computes \eqn{log( sum_i( exp(v[i] )) ) - log(n)} in a stable way.}
\usage{
logSumExpMean(v)
}
\description{
Computes \eqn{log( sum_i( exp(v[i] )) ) - log(n)} in a stable way.
}
\keyword{internal}
