% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_priors.R
\name{make_priors_and_init}
\alias{make_priors_and_init}
\title{Elicit priors and initialization from background dataset}
\usage{
make_priors_and_init(
  df.background,
  col.variables,
  col.item,
  use.priors = c("ML", "vague"),
  use.init = c("random", "vague"),
  ...
)
}
\arguments{
\item{df.background}{the background dataset}

\item{col.variables}{columns with variables: names or positions}

\item{col.item}{column with the id of the item (\eqn{i}): names or positions}

\item{use.priors}{see details}

\item{use.init}{see details}

\item{...}{additional variables for priors and init (see the description)}
}
\value{
a list of variables
}
\description{
Here we fix the hyperparameters for priors on \eqn{\theta_i}{theta_i} and \eqn{W_i}{W_i}, i.e., \eqn{B}, \eqn{U}, \eqn{\mu}{mu} and \eqn{\nu_w}{n_w}.
}
\details{
An appropriate initialization value for \eqn{W^{-1}_i}, \eqn{i=1,2} ($h_p$ and $h_d$ chains respectively) is also generated.

Notice that we have three chains in the LR computations, the same initialization is used thrice.
\subsection{Priors}{

\code{use.priors}:
\itemize{
\item \code{'ML'}: maximum likelihood estimation as in \insertCite{Bozza2008Probabilistic}{bayessource}
\item \code{'vague'}: low-information priors

\code{U} is \code{alpha*diag(p)}, B is \code{beta*diag(p)}, \code{mu} is \code{mu0}.

By default \verb{alpha = 1, beta = 1, mu0 = 0}
}

The Wishart dofs \code{nw} are set as small as possible without losing full rank:
\eqn{\nu_w = 2(p + 1) + 1}{n_w = 2*(p + 1) + 1}
}

\subsection{Initialization}{

Generate values for \eqn{W^{-1}_1}, \eqn{W^{-1}_2} that will be used as starting values for the Gibbs chains.
Two methods:

\code{use.init}:
\itemize{
\item \code{'random'}: initialize according to the model
\itemize{
\item generate one \eqn{W_i \sim IW_p(\nu_w, U)}{W_i ~ IW_p(n_w, U)}
\item then invert: \eqn{W^{-1}_i = (W_i)^{-1}}
}
\item \code{'vague'}: low-information initialization
\itemize{
\item \eqn{W_i = \alpha_{\text{init}} + \beta_{\text{init}} * \operatorname{diag}(p)}{W_i = alpha_init + beta_init * diag(p)}
}
}
}

\subsection{Parameters}{

Some constants can be changed by passing the new values to \code{...} :
\itemize{
\item \code{use.priors = 'vague'}: accepts parameters
\itemize{
\item \code{alpha} (default: \code{1})
\item \code{beta} (default: \code{1})
\item \code{mu0} (default: \code{0})
}
\item \code{use.init = 'vague'}: accepts parameters
\itemize{
\item \code{alpha_init} (default: \code{1})
\item \code{beta_init} (default: \code{100})
}
}
}

\subsection{Return values}{

A list of variables:
\itemize{
\item \code{mu}: the global mean
\item \code{B.inv}: the between-source prior covariance matrix, as the inverse
\item \code{U}: the Inverted Wishart scale matrix
\item \code{nw}: the Inverted Wishart dof \eqn{\nu_w}{n_w}
\item \code{W.inv.1}, \code{W.inv.2}: the initializations for the within-source covariance matrices, as their inverses
}
}
}
\examples{
# Use the iris data
head(iris, 3)
col_variables <- c(1,3)
col_item <- 5

# Elicitation using MLE
priors_init <- make_priors_and_init(
   df.background = iris,
   col.variables = col_variables,
   col.item = col_item
)

priors_init

priors_init_2 <- make_priors_and_init(
   df.background = iris,
   col.variables = col_variables,
   col.item = col_item,
   use.priors = "vague",
   alpha_init = 100
)

priors_init_2
}
\seealso{
Other core functions: 
\code{\link{bayessource-package}},
\code{\link{get_minimum_nw_IW}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{marginalLikelihood}()},
\code{\link{mcmc_postproc}()},
\code{\link{samesource_C}()},
\code{\link{two.level.multivariate.calculate.UC}()}
}
\concept{core functions}
