% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_postproc.R
\name{mcmc_postproc}
\alias{mcmc_postproc}
\title{Post-process Gibbs chain outputs.}
\usage{
mcmc_postproc(mcmc.output, compute.ML = FALSE, cumulative = TRUE)
}
\arguments{
\item{mcmc.output}{The coda object}

\item{compute.ML}{if TRUE, also compute the posterior means}

\item{cumulative}{if TRUE, also compute cumulative posterior means to assess precision}
}
\value{
a named list with multidimensional arrays for W, theta, and ML estimates
}
\description{
Extract and reshape MCMC samples from the posteriors for theta and W
.
Can be used to post-process outputs from \code{\link{marginalLikelihood}} when \code{output.mcmc} is TRUE.
Notice that it performs matrix inversions on every sample of W.
}
\details{
It is also possible to compute ML estimators for the posterior means, as well as the cumulated estimates.
The results will be appended to the returned list.
}
