% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Generate from multivariate normal.}
\usage{
rmvnorm(n, mu, Cov, is_chol = FALSE)
}
\arguments{
\item{n}{amount of samples to generate from}

\item{mu}{column vector for the mean}

\item{Cov}{covariance matrix}

\item{is_chol}{if TRUE, Cov is the upper Cholesky factor of Cov}
}
\value{
a nxp matrix of samples
}
\description{
Faster than \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}}. Implemented in C.
}
\seealso{
Other C++ functions: 
\code{\link{chol2inv}()},
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{isCholeskyOn}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rwish}()}

Other statistical functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dmvnorm}()},
\code{\link{dwishart}()},
\code{\link{riwish_Press}()},
\code{\link{rwish}()}
}
\concept{C++ functions}
\concept{statistical functions}
