% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwish}
\alias{rwish}
\title{Generate random sample from Wishart (faster).}
\usage{
rwish(v, S, is_chol = FALSE, return_chol = FALSE)
}
\arguments{
\item{v}{dof}

\item{S}{the scale matrix (pxp)}

\item{is_chol}{if TRUE, S is the upper Cholesky factor of S}

\item{return_chol}{if TRUE, the upper Cholesky factor is returned}
}
\value{
a single random variate from W(v, S)
}
\description{
Same code as \code{\link[stats]{rWishart}} function in package \pkg{stats}.
}
\section{Wishart parametrization}{


Uses Press/Anderson[Press2012,Anderson2003] parametrization.

\deqn{X ~ W(v, S)}
with \eqn{S = pxp} matrix, \eqn{v >= p} (the degrees of freedom).

Then:
\deqn{E[X] = v * S}
}

\references{
[Press2012] S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.

[Anderson2003] T. W. Anderson, An introduction to multivariate statistical analysis, 3rd ed. Hoboken, N.J: Wiley-Interscience, 2003.
}
