% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diwishart.R
\name{diwishart}
\alias{diwishart}
\title{Inverted Wishart density, parametrization according to Press.}
\usage{
diwishart(X, df, Sigma, log = FALSE, is.chol = FALSE)
}
\arguments{
\item{X}{observation}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{log}{if TRUE, return the log-density}

\item{is.chol}{if TRUE, Sigma and X are the upper Cholesky factors of Sigma and X}
}
\value{
the density in X
}
\description{
Implemented in R (slow).
}
\section{Inverted Wishart parametrization (Press)}{


Uses Press\link{Press2012} parametrization.

\deqn{X ~ IW(v, S)}
with \eqn{S = pxp} matrix, \eqn{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = S / (v - 2(p + 1))}
}

\references{
\link{Press2012} S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.
}
\seealso{
\code{\link{diwishart_inverse}}, \code{\link{dwishart}}
}
