% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diwishart_inverse_R.R
\name{diwishart_inverse_R}
\alias{diwishart_inverse_R}
\title{Inverted Wishart density from the inverse.}
\usage{
diwishart_inverse_R(X.inv, df, Sigma, log = FALSE, is.chol = FALSE)
}
\arguments{
\item{X.inv}{inverse of X (the observation)}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{log}{if TRUE, return the log-density}

\item{is.chol}{if TRUE, Sigma and X.inv are the upper Cholesky factors of Sigma and X.inv}
}
\value{
the density in X
}
\description{
Computes the density of an Inverted Wishart (df, Sigma) in X, by supplying (X^(-1), df, Sigma) rather than (X, df, Sigma).
Avoids a matrix inversion.
}
\details{
Computes the pdf p_X(x) by knowing x^(-1)
}
\section{Inverted Wishart parametrization (Press)}{


Uses \insertCite{Press2012Applied}{bayessource} parametrization.

\deqn{X \sim IW(\nu, S)}{X ~ IW(v, S)}
with \eqn{S} is a \eqn{p \times p}{p x p} matrix, \eqn{\nu > 2p}{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = \frac{S}{\nu - 2(p + 1)}}{E[X] = S/(n - 2(p + 1))}
}

\references{
\insertAllCited{}
}
\seealso{
\code{\link{diwishart_inverse}}, \code{\link{diwishart}}

Other R functions: 
\code{\link{diwishart}()},
\code{\link{dwishart}()},
\code{\link{riwish_Press}()}

Other statistical functions: 
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dmvnorm}()},
\code{\link{dwishart}()},
\code{\link{riwish_Press}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other Wishart functions: 
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dwishart}()},
\code{\link{get_minimum_nw_IW}()},
\code{\link{riwish_Press}()},
\code{\link{rwish}()}
}
\concept{R functions}
\concept{Wishart functions}
\concept{statistical functions}
\keyword{internal}
