% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwishart.R
\name{dwishart}
\alias{dwishart}
\title{Wishart density.}
\usage{
dwishart(X, df, Sigma, log = FALSE, is.chol = FALSE)
}
\arguments{
\item{X}{the observation}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{log}{if TRUE, return the log-density}

\item{is.chol}{if TRUE, Sigma and X are the upper Cholesky factors of Sigma and X}
}
\value{
the density in X
}
\description{
Density of X ~ Wishart(df, Sigma).
}
\section{Wishart parametrization}{


Uses Press/Anderson\link{Press2012,Anderson2003} parametrization.

\deqn{X ~ W(v, S)}
with \eqn{S = pxp} matrix, \eqn{v >= p} (the degrees of freedom).

Then:
\deqn{E[X] = v * S}
}

\references{
\link{Press2012} S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.

\link{Anderson2003} T. W. Anderson, An introduction to multivariate statistical analysis, 3rd ed. Hoboken, N.J: Wiley-Interscience, 2003.
}
