% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwishart.R
\name{dwishart}
\alias{dwishart}
\title{Wishart density.}
\usage{
dwishart(X, df, Sigma, log = FALSE, is.chol = FALSE)
}
\arguments{
\item{X}{the observation}

\item{df}{degrees of freedom}

\item{Sigma}{scale matrix}

\item{log}{if TRUE, return the log-density}

\item{is.chol}{if TRUE, Sigma and X are the upper Cholesky factors of Sigma and X}
}
\value{
the density in X
}
\description{
Density of X ~ Wishart(df, Sigma).
}
\section{Wishart parametrization}{


Uses \insertCite{Press2012Applied,Anderson2003introduction}{bayessource} parametrization.

\deqn{X \sim W(\nu, S)}{X ~ W(v, S)}
with \eqn{S} is a \eqn{p \times p}{pxp} matrix, \eqn{\nu >= p}{v >= p} (the degrees of freedom).

Then:
\deqn{E[X] = \nu S}{E[X] = v * S}
}

\references{
\insertAllCited{}
}
\seealso{
Other R functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart}()},
\code{\link{riwish_Press}()}

Other statistical functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dmvnorm}()},
\code{\link{riwish_Press}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other Wishart functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{get_minimum_nw_IW}()},
\code{\link{riwish_Press}()},
\code{\link{rwish}()}
}
\concept{R functions}
\concept{Wishart functions}
\concept{statistical functions}
