% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_priors.R
\name{get_minimum_nw_IW}
\alias{get_minimum_nw_IW}
\title{Get minimum degrees of freedom for Inverted Wishart}
\usage{
get_minimum_nw_IW(p)
}
\arguments{
\item{p}{number of variables}
}
\value{
minimum dof \eqn{\nu_w}{n_w}
}
\description{
Returns minimum value for degrees of freedom such that the Inverted Wishart has a well-defined expected value.
}
\details{
Uses \insertCite{Press2012Applied}{bayessource} parametrization.

\deqn{X \sim IW(\nu, S)}{X ~ IW(n_w, S)}

with \eqn{S = pxp} matrix, \eqn{n_w > 2p}{\nu_w > 2p} (the degrees of freedom).

Then:

\deqn{E[X] = S / (\nu_w - 2(p + 1))}{E[X] = S / (n_w - 2(p + 1))}

Finally, the minimum dof \eqn{\nu_w}{n_w} are: \eqn{\nu_w = 2(p + 1) + 1}{n_w = 2*(p + 1) + 1}.
}
\references{
\insertAllCited{}
}
\seealso{
Other core functions: 
\code{\link{bayessource-package}},
\code{\link{make_priors_and_init}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{marginalLikelihood}()},
\code{\link{mcmc_postproc}()},
\code{\link{samesource_C}()},
\code{\link{two.level.multivariate.calculate.UC}()}

Other Wishart functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dwishart}()},
\code{\link{riwish_Press}()},
\code{\link{rwish}()}
}
\concept{Wishart functions}
\concept{core functions}
