% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_priors.R
\name{get_minimum_nw_IW}
\alias{get_minimum_nw_IW}
\title{Get minimum degrees of freedom for Inverted Wishart}
\usage{
get_minimum_nw_IW(p)
}
\arguments{
\item{p}{dimension}
}
\value{
minimum dof
}
\description{
Returns minimum value for degrees of freedom such that Inverted Wishart has a mean..
}
\details{
Uses Press\link{1} parametrization.

\deqn{X ~ IW(nw, S)}, with \eqn{S = pxp} matrix, \eqn{nw > 2p} (the degrees of freedom).
Then:
\deqn{E[X] = S / (nw - 2(p + 1))}

Finally, the minimum dof $v$ is $ nw = 2*(p + 1) + 1 $.
}
\references{
\link{1}S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.
}
