% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_sympd_tol}
\alias{inv_sympd_tol}
\title{Compute the inverse of a symmetric positive definite matrix}
\usage{
inv_sympd_tol(U_sympd)
}
\description{
Compute the inverse of a symmetric positive definite matrix.
Does not output warnings on default symmetry tolerance: basically, symmetry is forced.
}
\references{
https://github.com/RcppCore/RcppArmadillo/issues/257
}
\seealso{
Other C++ functions: 
\code{\link{chol2inv}()},
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_triangular}()},
\code{\link{isCholeskyOn}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other math functions: 
\code{\link{chol2inv}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_pd}()},
\code{\link{inv_triangular}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()}
}
\concept{C++ functions}
\concept{math functions}
\keyword{internal}
