% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_triangular}
\alias{inv_triangular}
\title{Upper triangular matrix inversion}
\usage{
inv_triangular(U)
}
\description{
Quickly computes the inverse of a upper triangular matrix (e.g. a Cholesky factor).
}
\details{
Equivalent R code:

\code{X.chol.inv <- backsolve(r = X.chol, x = diag(p))}
}
\keyword{internal}
