% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ldet_from_Cholesky}
\alias{ldet_from_Cholesky}
\title{log-determinant from Cholesky factor}
\usage{
ldet_from_Cholesky(T_chol)
}
\description{
If \eqn{A = U' U}, compute log(det(A)) from U
}
\seealso{
Other C++ functions: 
\code{\link{chol2inv}()},
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{isCholeskyOn}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other math functions: 
\code{\link{chol2inv}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_pd}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()}
}
\concept{C++ functions}
\concept{math functions}
\keyword{internal}
