% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logCumsumExp}
\alias{logCumsumExp}
\title{Computes log( cumsum_i( exp(v\link{i}) ) ) in a stable way.}
\usage{
logCumsumExp(v)
}
\description{
Computes log( cumsum_i( exp(v\link{i}) ) ) in a stable way.
}
\keyword{internal}
