% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_priors.R
\name{make_priors_and_init}
\alias{make_priors_and_init}
\title{Elicit priors and initialization from background dataset}
\usage{
make_priors_and_init(
  df.background,
  col.variables,
  col.item,
  use.priors = "ML",
  use.init = "random",
  ...
)
}
\arguments{
\item{df.background}{the background dataset}

\item{col.variables}{columns with variables}

\item{col.item}{column with item id}

\item{use.priors}{see details}

\item{use.init}{see details}

\item{...}{additional variables for priors, init}
}
\value{
a list of variables
}
\description{
Here we fix the hyperparameters for priors on \eqn{theta_i} and \eqn{W_i}, i.e., \eqn{B}, \eqn{U}, \eqn{mu} and \eqn{n_w}.
}
\details{
Here we fix the hyperparameters for priors on \eqn{theta_i} and \eqn{W_i}, i.e., \eqn{B}, \eqn{U}, \eqn{mu} and \eqn{n_w}.

An appropriate initialization value for \eqn{W^{-1}_i}, \eqn{i=1,2} is also generated.

Notice that we have three chains in the LR computations, the same initialization is used thrice.
}
\section{Priors}{


\code{use.priors}:
\itemize{
\item \code{'ML'}: maximum likelihood estimation
\item \code{'vague'}: low-information priors

\code{U} is \code{alpha*diag(p)}, B is \code{beta*diag(p)}, \code{mu} is \code{mu0}.
By default \verb{alpha = 1, beta = 1, mu0 = 0}
}

The Wishart dofs \code{nw} are set as small as possible without losing full rank:
\eqn{nw = 2*(p + 1) + 1}
}

\section{Initialization}{


\code{use.init}:
\itemize{
\item \code{'random'}: initialize according to the model
\item \code{'vague'}: low-information initialization
\eqn{W_i} is \code{alpha_init + beta_init *diag(p)}
}

Some constants can be changed by passing the new values to \code{...} :
\itemize{
\item \code{use.priors = 'vague'}:
\verb{alpha = 1, beta = 1, mu0 = 0}
\item \code{use.init = 'vague'}:
\verb{alpha_init = 1, beta_init = 100}
}
}

\section{Returns}{


A list of variables:
\itemize{
\item \code{mu}: the global mean
\item \code{B.inv}: the between prior covariance, as the inverse
\item \code{U}: the Inverted Wishart scale matrix
\item \code{nw}: the Inverted Wishart dof
\item \code{W.inv.1}, \code{W.inv.2}: the within cov initializations, as inverses
}
}

