% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samesource_cpp.R
\name{marginalLikelihood}
\alias{marginalLikelihood}
\title{Fast Bayesian marginal likelihood for the Normal - Inverted Wishart model.}
\usage{
marginalLikelihood(X, n.iter, B.inv, W.inv, U, nw, mu, burn.in,
  output.mcmc = FALSE, verbose = FALSE, Gibbs_only = FALSE)
}
\arguments{
\item{X}{the dataset}

\item{n.iter}{number of MC iterations}

\item{B.inv}{prior inverse of between covariance matrix}

\item{W.inv}{prior inverse of within covariance matrix}

\item{U}{covariance matrix for the mean}

\item{nw}{degrees of freedom}

\item{mu}{prior mean}

\item{burn.in}{burn-in iterations}

\item{output.mcmc}{output the entire chain}

\item{verbose}{if TRUE, be verbose}

\item{Gibbs_only}{if TRUE, only return the Gibbs posterior samples. Implies chain_output = TRUE.}
}
\value{
the log-marginal likelihood value, or a list(the log-ml value, coda object with the posterior samples)
}
\description{
Implemented in C.
See \code{\link{diwishart_inverse}} for the parametrization of the Inverted Wishart.
See \code{\link[bayessource]{marginalLikelihood_internal}} for further documentation.
}
\section{Normal-Inverse-Wishart model[Bozza2008]}{


Observation level:
\deqn{X_{ij} ~ N(theta_i, W_i)} (i = population, j = items in population)

Group level:
\deqn{theta_i ~ N(\mu, B)}
\deqn{W_i ~ IW(nw, U)}

Hyperparameters:
\eqn{B, U, nw, \mu}

A Gibbs sampler supplies theta, W^{(-1)}.
}

\section{Inverted Wishart parametrization (Press)}{


Uses Press[Press2012] parametrization.

\deqn{X ~ IW(v, S)}
with \eqn{S = pxp} matrix, \eqn{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = S / (v - 2(p + 1))}
}

\references{
[Bozza2008] S. Bozza, F. Taroni, R. Marquis, and M. Schmittbuhl, “Probabilistic evaluation of handwriting evidence: likelihood ratio for authorship,” Journal of the Royal Statistical Society: Series C (Applied Statistics), vol. 57, no. 3, pp. 329–341, Jun. 2008.

[Press2012] S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.
}
\seealso{
marginalLikelihood_internal
}
