% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginalLikelihood_internal}
\alias{marginalLikelihood_internal}
\title{Fast Bayesian same source hypothesis. Gaussian MV.
To be called by the R wrapper.}
\usage{
marginalLikelihood_internal(X, n_iter, B_inv, W_inv, U, nw, mu, burn_in,
  chain_output = FALSE, verbose = FALSE, Gibbs_only = FALSE)
}
\arguments{
\item{X}{the dataset}

\item{n_iter}{number of MC iterations}

\item{B_inv}{prior inverse of between covariance matrix}

\item{W_inv}{initialization for prior inverse of within covariance matrix}

\item{U}{covariance matrix for the mean}

\item{nw}{degrees of freedom}

\item{mu}{prior mean}

\item{burn_in}{burn-in iterations}

\item{chain_output}{if true, output the entire chain as a list (ML-value, samples from theta, samples from W_inv)}

\item{verbose}{if TRUE, be verbose}

\item{Gibbs_only}{if TRUE, only return the Gibbs posterior samples. Implies chain_output = TRUE.}
}
\description{
Fast Bayesian same source hypothesis. Gaussian MV.
To be called by the R wrapper.
}
\section{Normal-Inverse-Wishart model[Bozza2008]}{


Observation level:
\deqn{X_{ij} ~ N(theta_i, W_i)} (i = population, j = items in population)

Group level:
\deqn{theta_i ~ N(\mu, B)}
\deqn{W_i ~ IW(nw, U)}

Hyperparameters:
\eqn{B, U, nw, \mu}

A Gibbs sampler supplies theta, W^{(-1)}.
}

\references{
[Bozza2008] S. Bozza, F. Taroni, R. Marquis, and M. Schmittbuhl, “Probabilistic evaluation of handwriting evidence: likelihood ratio for authorship,” Journal of the Royal Statistical Society: Series C (Applied Statistics), vol. 57, no. 3, pp. 329–341, Jun. 2008.
}
\keyword{internal}
