% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_priors.R
\name{riwish_Press}
\alias{riwish_Press}
\title{Generate random sample from Inverted Wishart.}
\usage{
riwish_Press(v, S)
}
\arguments{
\item{v}{dof (\eqn{> 2p})}

\item{S}{the scale matrix (pxp)}
}
\value{
a single random variate from IW(v, S)
}
\description{
Using Press parametrization.
}
\details{
Uses \pkg{MCMCpack}::riwish.
}
\section{Inverted Wishart parametrization (Press)}{


Uses Press[Press2012] parametrization.

\deqn{X ~ IW(v, S)}
with \eqn{S = pxp} matrix, \eqn{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = S / (v - 2(p + 1))}
}

\references{
[Press2012] S. J. Press, Applied multivariate analysis: using Bayesian and frequentist methods of inference. Courier Corporation, 2012.
}
