% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_priors.R
\name{riwish_Press}
\alias{riwish_Press}
\title{Generate random sample from Inverted Wishart.}
\usage{
riwish_Press(v, S)
}
\arguments{
\item{v}{dof (\eqn{> 2p})}

\item{S}{the scale matrix (pxp)}
}
\value{
a single random variate from IW(v, S)
}
\description{
Using \insertCite{Press2012Applied}{bayessource} parametrization.
}
\section{Inverted Wishart parametrization (Press)}{


Uses \insertCite{Press2012Applied}{bayessource} parametrization.

\deqn{X \sim IW(\nu, S)}{X ~ IW(v, S)}
with \eqn{S} is a \eqn{p \times p}{p x p} matrix, \eqn{\nu > 2p}{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = \frac{S}{\nu - 2(p + 1)}}{E[X] = S/(n - 2(p + 1))}
}

\references{
\insertAllCited{}
}
\seealso{
Other R functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart}()},
\code{\link{dwishart}()}

Other statistical functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dmvnorm}()},
\code{\link{dwishart}()},
\code{\link{rmvnorm}()},
\code{\link{rwish}()}

Other Wishart functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dwishart}()},
\code{\link{get_minimum_nw_IW}()},
\code{\link{rwish}()}
}
\concept{R functions}
\concept{Wishart functions}
\concept{statistical functions}
