% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Generate from multivariate normal.}
\usage{
rmvnorm(n, mu, Cov, is_chol = FALSE)
}
\arguments{
\item{n}{amount of samples to generate from}

\item{mu}{column vector for the mean}

\item{Cov}{covariance matrix}

\item{is_chol}{if TRUE, Cov is the upper Cholesky factor of Cov}
}
\value{
a nxp matrix of samples
}
\description{
Faster than \code{rmvnorm} in package \pkg{mvtnorm}. Implemented in C.
}
