% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwish}
\alias{rwish}
\title{Generate random sample from Wishart (faster).}
\usage{
rwish(v, S, is_chol = FALSE, return_chol = FALSE)
}
\arguments{
\item{v}{dof}

\item{S}{the scale matrix (pxp)}

\item{is_chol}{if TRUE, S is the upper Cholesky factor of S}

\item{return_chol}{if TRUE, the upper Cholesky factor is returned}
}
\value{
a single random variate from W(v, S)
}
\description{
Same code as \code{\link[stats]{rWishart}} function in package \pkg{stats}.
}
\section{Wishart parametrization}{


Uses \insertCite{Press2012Applied,Anderson2003introduction}{bayessource} parametrization.

\deqn{X \sim W(\nu, S)}{X ~ W(\nu, S)}
with \eqn{S} is a \eqn{p \times p}{pxp} matrix, \eqn{\nu >= p}{v >= p} (the degrees of freedom).

Then:
\deqn{E[X] = \nu S}{E[X] = v * S}
}

\references{
\insertAllCited{}
}
\seealso{
Other C++ functions: 
\code{\link{chol2inv}()},
\code{\link{diwishart_inverse}()},
\code{\link{dmvnorm}()},
\code{\link{inv_Cholesky_from_Cholesky}()},
\code{\link{inv_sympd_tol}()},
\code{\link{inv_triangular}()},
\code{\link{isCholeskyOn}()},
\code{\link{ldet_from_Cholesky}()},
\code{\link{logCummeanExp}()},
\code{\link{logCumsumExp}()},
\code{\link{logSumExpMean}()},
\code{\link{logSumExp}()},
\code{\link{marginalLikelihood_internal}()},
\code{\link{rmvnorm}()}

Other statistical functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dmvnorm}()},
\code{\link{dwishart}()},
\code{\link{riwish_Press}()},
\code{\link{rmvnorm}()}

Other Wishart functions: 
\code{\link{diwishart_inverse_R}()},
\code{\link{diwishart_inverse}()},
\code{\link{diwishart}()},
\code{\link{dwishart}()},
\code{\link{get_minimum_nw_IW}()},
\code{\link{riwish_Press}()}
}
\concept{C++ functions}
\concept{Wishart functions}
\concept{statistical functions}
