% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samesource_cpp.R
\name{samesource_C}
\alias{samesource_C}
\title{Fast Bayesian same source hypothesis for the Normal - Inverted Wishart model.}
\usage{
samesource_C(
  quest,
  ref,
  n.iter,
  B.inv,
  W.inv.1,
  W.inv.2,
  U,
  nw,
  mu,
  burn.in,
  verbose = FALSE,
  marginals = FALSE
)
}
\arguments{
\item{quest}{the questioned dataset}

\item{ref}{the reference dataset}

\item{n.iter}{number of MC iterations}

\item{B.inv}{prior inverse of between covariance matrix}

\item{W.inv.1}{prior inverse of within covariance matrix (questioned items)}

\item{W.inv.2}{prior inverse of within covariance matrix (reference items)}

\item{U}{covariance matrix for the mean}

\item{nw}{degrees of freedom}

\item{mu}{prior mean}

\item{burn.in}{burn-in iterations}

\item{verbose}{if TRUE, be verbose}

\item{marginals}{return the marginal likelihoods in the LR formula (default: FALSE)}
}
\value{
the log-BF value (base e), or a list with the log-BF and the marginals: \code{list(value, log_ml_Hp, log_ml_Hd_ref, log_ml_Hd_quest)}
}
\description{
Implemented in C.
See \code{\link[bayessource]{diwishart_inverse}} for the parametrization of the Inverted Wishart.
See \code{\link[bayessource]{marginalLikelihood_internal}} for further documentation.
}
\section{Normal-Inverse-Wishart model\insertCite{Bozza2008Probabilistic}{bayessource}}{


Observation level:
\deqn{X_{ij} ~ N_p(theta_i, W_i)} (i = population, j = items in population)

Group level:
\deqn{theta_i ~ N_p(\mu, B)}
\deqn{W_i ~ IW_p(nw, U)}

Hyperparameters:
\eqn{B, U, nw, \mu}

A Gibbs sampler supplies theta, W^{(-1)}.
}

\section{Inverted Wishart parametrization (Press)}{


Uses \insertCite{Press2012Applied}{bayessource} parametrization.

\deqn{X ~ IW(v, S)}
with \eqn{S = pxp} matrix, \eqn{v > 2p} (the degrees of freedom).

Then:
\deqn{E[X] = S / (v - 2(p + 1))}
}

\references{
\insertAllCited{}
}
\seealso{
marginalLikelihood
}
