% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_level_functions.R
\name{two.level.multivariate.calculate.UC}
\alias{two.level.multivariate.calculate.UC}
\title{Calculate Between/Within covariance matrices}
\usage{
two.level.multivariate.calculate.UC(
  population,
  idx.variables,
  idx.item,
  singular.action = "skip"
)
}
\arguments{
\item{population}{matrix or dataframe containing data columns, item column and other columns}

\item{idx.variables}{data columns (indexes or character vectors)}

\item{idx.item}{item column (index or character vector)}

\item{singular.action}{action to take when one of the items is observed once; one of \code{'skip'} (default) or \code{'fail'}}
}
\value{
list with items:
\itemize{
\item \code{group.means} (matrix, row*variable)
\item \code{all.means} (vector)
\item \code{W} (the within covariance matrix)
\item \code{B} (the between covariance matrix)
}
}
\description{
Two-level function: item is chosen with the third parameter.
}
\details{
Items who are observed only once cannot contribute to within covariance estimation.
In that case, either we skip its contribute (default) or fail.

Multivariate functions supporting paper:
Bozza, A., Taroni, F., Marquis, R., M. Schmittbuhl
}
\keyword{internal}
