% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_level_functions.R
\name{two.level.multivariate.calculate.UC}
\alias{two.level.multivariate.calculate.UC}
\title{Calculate Between/Within covariance matrices}
\usage{
two.level.multivariate.calculate.UC(population, idx.variables, idx.item)
}
\arguments{
\item{population}{matrix or dataframe containing data columns, item column and other columns}

\item{idx.item}{item column (index or character vector)}

\item{variables}{data columns (indexes or character vectors)}
}
\value{
list with items:
- group.means (matrix, row*variable)
- all.means (vector)
- W (the within covariance matrix)
- B (the between covariance matrix)
}
\description{
Two-level function: item is chosen with the third parameter.
}
\details{
Multivariate functions supporting paper:
Bozza, A., Taroni, F., Marquis, R., M. Schmittbuhl

This function is RNG-free.
}
